+++ Rubeus/lib/LSA.cs
@@ -8,6 +8,7 @@ using System.Security;
 using System.Reflection;
 using System.Security.AccessControl;
 using Microsoft.Win32;
+using ConsoleTables;
 
 namespace Rubeus
 {
@@ -516,7 +517,7 @@ namespace Rubeus
                                     }
                                     Console.WriteLine("\r\n  UserName                 : {0}", username);
                                     Console.WriteLine("  Domain                   : {0}", domain);
-                                    Console.WriteLine("  LogonId                  : {0}", data.LoginID.LowPart);
+                                    Console.WriteLine("  LogonId                  : {0}", data.LoginID);
                                     Console.WriteLine("  UserSID                  : {0}", sid.Value);
                                     Console.WriteLine("  AuthenticationPackage    : {0}", authpackage);
                                     Console.WriteLine("  LogonType                : {0}", logonType);
@@ -896,7 +897,7 @@ namespace Rubeus
                                 {
                                     Console.WriteLine("\r\n  UserName                 : {0}", username);
                                     Console.WriteLine("  Domain                   : {0}", domain);
-                                    Console.WriteLine("  LogonId                  : {0}", data.LoginID.LowPart);
+                                    Console.WriteLine("  LogonId                  : {0}", data.LoginID);
                                     Console.WriteLine("  UserSID                  : {0}", sid.Value);
                                     Console.WriteLine("  AuthenticationPackage    : {0}", authpackage);
                                     Console.WriteLine("  LogonType                : {0}", logonType);
@@ -1317,6 +1318,200 @@ namespace Rubeus
             // disconnect from LSA
             Interop.LsaDeregisterLogonProcess(lsaHandle);
         }
+
+        public static void TriageKerberosTickets(Interop.LUID targetLuid, string user = "", string service = "")
+        {
+            // lists Kerberos tickets for all users on the system (assuming elevation)
+
+            //  first elevates to SYSTEM and uses LsaRegisterLogonProcessHelper connect to LSA
+            //  then calls LsaCallAuthenticationPackage w/ a KerbQueryTicketCacheMessage message type to enumerate all cached tickets
+
+            // adapted partially from Vincent LE TOUX' work
+            //      https://github.com/vletoux/MakeMeEnterpriseAdmin/blob/master/MakeMeEnterpriseAdmin.ps1#L2939-L2950
+            // and https://www.dreamincode.net/forums/topic/135033-increment-memory-pointer-issue/
+            // also Jared Atkinson's work at https://github.com/Invoke-IR/ACE/blob/master/ACE-Management/PS-ACE/Scripts/ACE_Get-KerberosTicketCache.ps1
+
+            Console.WriteLine("\r\n\r\n[*] Action: Triage Kerberos Tickets\r\n");
+
+            if (!Helpers.IsHighIntegrity())
+            {
+                Console.WriteLine("[X] You need to have an elevated context to dump other users' Kerberos tickets :( \r\n");
+                return;
+            }
+
+            if ((ulong)targetLuid != 0)
+            {
+                Console.WriteLine("[*] Target LUID     : 0x{0:x}", (ulong)targetLuid);
+            }
+            if (!String.IsNullOrEmpty(user))
+            {
+                Console.WriteLine("[*] Target user     : {0}", user);
+            }
+            if (!String.IsNullOrEmpty(service))
+            {
+                Console.WriteLine("[*] Target service  : {0}", service);
+            }
+            if (((ulong)targetLuid != 0) || (!String.IsNullOrEmpty(user)) || (!String.IsNullOrEmpty(service)))
+            {
+                Console.WriteLine();
+            }
+
+            var table = new ConsoleTable("LUID", "UserName", "Service", "EndTime");
+            int retCode;
+            int authPack;
+            string name = "kerberos";
+            Interop.LSA_STRING_IN LSAString;
+            LSAString.Length = (ushort)name.Length;
+            LSAString.MaximumLength = (ushort)(name.Length + 1);
+            LSAString.Buffer = name;
+
+            IntPtr lsaHandle = LsaRegisterLogonProcessHelper();
+
+            // if the original call fails then it is likely we don't have SeTcbPrivilege
+            // to get SeTcbPrivilege we can Impersonate a NT AUTHORITY\SYSTEM Token
+            if (lsaHandle == IntPtr.Zero)
+            {
+                string currentName = System.Security.Principal.WindowsIdentity.GetCurrent().Name;
+                if (currentName == "NT AUTHORITY\\SYSTEM")
+                {
+                    // if we're already SYSTEM, we have the proper privilegess to get a Handle to LSA with LsaRegisterLogonProcessHelper
+                    lsaHandle = LsaRegisterLogonProcessHelper();
+                }
+                else
+                {
+                    // elevated but not system, so gotta GetSystem() first
+                    Helpers.GetSystem();
+                    // should now have the proper privileges to get a Handle to LSA
+                    lsaHandle = LsaRegisterLogonProcessHelper();
+                    // we don't need our NT AUTHORITY\SYSTEM Token anymore so we can revert to our original token
+                    Interop.RevertToSelf();
+                }
+            }
+
+            try
+            {
+                // obtains the unique identifier for the kerberos authentication package.
+                retCode = Interop.LsaLookupAuthenticationPackage(lsaHandle, ref LSAString, out authPack);
+
+                // first return all the logon sessions
+                DateTime systime = new DateTime(1601, 1, 1, 0, 0, 0, 0); //win32 systemdate
+                UInt64 count;
+                IntPtr luidPtr = IntPtr.Zero;
+                IntPtr iter = luidPtr;
+
+                uint ret = Interop.LsaEnumerateLogonSessions(out count, out luidPtr);  // get an array of pointers to LUIDs
+
+                for (ulong i = 0; i < count; i++)
+                {
+                    IntPtr sessionData;
+                    ret = Interop.LsaGetLogonSessionData(luidPtr, out sessionData);
+                    Interop.SECURITY_LOGON_SESSION_DATA data = (Interop.SECURITY_LOGON_SESSION_DATA)Marshal.PtrToStructure(sessionData, typeof(Interop.SECURITY_LOGON_SESSION_DATA));
+
+                    // if we have a valid logon
+                    if (data.PSiD != IntPtr.Zero)
+                    {
+                        // user session data
+                        string username = Marshal.PtrToStringUni(data.Username.Buffer).Trim();
+                        System.Security.Principal.SecurityIdentifier sid = new System.Security.Principal.SecurityIdentifier(data.PSiD);
+                        string domain = Marshal.PtrToStringUni(data.LoginDomain.Buffer).Trim();
+                        string authpackage = Marshal.PtrToStringUni(data.AuthenticationPackage.Buffer).Trim();
+                        Interop.SECURITY_LOGON_TYPE logonType = (Interop.SECURITY_LOGON_TYPE)data.LogonType;
+                        DateTime logonTime = systime.AddTicks((long)data.LoginTime);
+                        string logonServer = Marshal.PtrToStringUni(data.LogonServer.Buffer).Trim();
+                        string dnsDomainName = Marshal.PtrToStringUni(data.DnsDomainName.Buffer).Trim();
+                        string upn = Marshal.PtrToStringUni(data.Upn.Buffer).Trim();
+
+                        IntPtr ticketsPointer = IntPtr.Zero;
+                        DateTime sysTime = new DateTime(1601, 1, 1, 0, 0, 0, 0);
+
+                        int returnBufferLength = 0;
+                        int protocalStatus = 0;
+
+                        Interop.KERB_QUERY_TKT_CACHE_REQUEST tQuery = new Interop.KERB_QUERY_TKT_CACHE_REQUEST();
+                        Interop.KERB_QUERY_TKT_CACHE_RESPONSE tickets = new Interop.KERB_QUERY_TKT_CACHE_RESPONSE();
+                        Interop.KERB_TICKET_CACHE_INFO_EX ticket;
+
+                        // input object for querying the ticket cache for a specific logon ID
+                        Interop.LUID userLogonID = new Interop.LUID(data.LoginID);
+                        tQuery.LogonId = userLogonID;
+
+                        if (((ulong)targetLuid == 0) || (data.LoginID == targetLuid))
+                        {
+                            tQuery.MessageType = Interop.KERB_PROTOCOL_MESSAGE_TYPE.KerbQueryTicketCacheExMessage;
+
+                            // query LSA, specifying we want the ticket cache
+                            IntPtr tQueryPtr = Marshal.AllocHGlobal(Marshal.SizeOf(tQuery));
+                            Marshal.StructureToPtr(tQuery, tQueryPtr, false);
+                            retCode = Interop.LsaCallAuthenticationPackage(lsaHandle, authPack, tQueryPtr, Marshal.SizeOf(tQuery), out ticketsPointer, out returnBufferLength, out protocalStatus);
+
+                            if (ticketsPointer != IntPtr.Zero)
+                            {
+                                // parse the returned pointer into our initial KERB_QUERY_TKT_CACHE_RESPONSE structure
+                                tickets = (Interop.KERB_QUERY_TKT_CACHE_RESPONSE)Marshal.PtrToStructure((System.IntPtr)ticketsPointer, typeof(Interop.KERB_QUERY_TKT_CACHE_RESPONSE));
+                                int count2 = tickets.CountOfTickets;
+
+                                if (count2 != 0)
+                                {
+                                    // get the size of the structures we're iterating over
+                                    Int32 dataSize = Marshal.SizeOf(typeof(Interop.KERB_TICKET_CACHE_INFO_EX));
+
+                                    for (int j = 0; j < count2; j++)
+                                    {
+                                        // iterate through the result structures
+                                        IntPtr currTicketPtr = (IntPtr)(long)((ticketsPointer.ToInt64() + (int)(8 + j * dataSize)));
+
+                                        // parse the new ptr to the appropriate structure
+                                        ticket = (Interop.KERB_TICKET_CACHE_INFO_EX)Marshal.PtrToStructure(currTicketPtr, typeof(Interop.KERB_TICKET_CACHE_INFO_EX));
+
+                                        DateTime startTime = DateTime.FromFileTime(ticket.StartTime);
+                                        DateTime endTime = DateTime.FromFileTime(ticket.EndTime);
+                                        DateTime renewTime = DateTime.FromFileTime(ticket.RenewTime);
+
+                                        string ticketFlags = ((Interop.TicketFlags)ticket.TicketFlags).ToString();
+
+                                        // extract the server name/realm and client name/realm
+                                        string serverName = Marshal.PtrToStringUni(ticket.ServerName.Buffer, ticket.ServerName.Length / 2);
+                                        string serverRealm = Marshal.PtrToStringUni(ticket.ServerRealm.Buffer, ticket.ServerRealm.Length / 2);
+                                        string clientName = Marshal.PtrToStringUni(ticket.ClientName.Buffer, ticket.ClientName.Length / 2);
+                                        string clientRealm = Marshal.PtrToStringUni(ticket.ClientRealm.Buffer, ticket.ClientRealm.Length / 2);
+
+                                        if (String.IsNullOrEmpty(user) || (Regex.IsMatch(clientName, user, RegexOptions.IgnoreCase)))
+                                        {
+                                            if (String.IsNullOrEmpty(service) || (Regex.IsMatch(serverName, service, RegexOptions.IgnoreCase)))
+                                            {
+                                                table.AddRow(data.LoginID.ToString(), String.Format("{0} @ {1}", clientName, clientRealm), serverName, endTime.ToString());
+                                            }
+                                        }
+                                    }
+                                }
+                            }
+
+                            // cleanup
+                            Interop.LsaFreeReturnBuffer(ticketsPointer);
+                            Marshal.FreeHGlobal(tQueryPtr);
+                        }
+                    }
+
+                    // move the pointer forward
+                    luidPtr = (IntPtr)((long)luidPtr.ToInt64() + Marshal.SizeOf(typeof(Interop.LUID)));
+
+                    // cleaup
+                    Interop.LsaFreeReturnBuffer(sessionData);
+                }
+                Interop.LsaFreeReturnBuffer(luidPtr);
+
+                // disconnect from LSA
+                Interop.LsaDeregisterLogonProcess(lsaHandle);
+
+                // write out the table
+                table.Write();
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine("[X] Exception: {0}", ex);
+            }
+        }
+
 
         public static List<KRB_CRED> ExtractTGTs(Interop.LUID targetLuid, bool includeComputerAccounts = false)
         {