+++ CHANGELOG.md
@@ -4,7 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.3.2] - 2019-02-05
+## [1.3.3] - 2019-02-07
+### Added
+* **triage** action
+    * Quickly triages the users and present tickets on a machine
+
+### Changed
+* **dump** and **klist** changed default LUID output to hex format
+
+
+## [1.3.2] - 2019-02-06
 ### Added
 * **kerberoast** and **asreproast** actions
     * Added /outfile:X to output hashes to a file, one hash per line